[runtime] move all str related function to string.h/string.cc
