+++ runtime/core/bin/decoder_main.cc
@@ -11,6 +11,7 @@
 #include "frontend/wav.h"
 #include "utils/flags.h"
 #include "utils/log.h"
+#include "utils/string.h"
 #include "utils/timer.h"
 #include "utils/utils.h"
 
@@ -69,8 +70,9 @@ int main(int argc, char *argv[]) {
     wenet::TorchAsrDecoder decoder(feature_pipeline, model, symbol_table,
                                    *decode_config, fst);
 
-    int wave_dur = static_cast<int>(static_cast<float>(
-                wav_reader.num_sample()) / wav_reader.sample_rate() * 1000);
+    int wave_dur =
+        static_cast<int>(static_cast<float>(wav_reader.num_sample()) /
+                         wav_reader.sample_rate() * 1000);
     int decode_time = 0;
     while (true) {
       wenet::Timer timer;
